requirements.txt
fastapi==0.135.3
psycopg==3.3.3
uvicorn==0.42.0
python-dotenv==1.2.2
pydantic==2.12.5
passlib==1.7.4
python-jose==3.5.0